@@ -153,34 +154,38 @@
     "@redux-devtools/cli": "^1.0.6",
     "@testing-library/react-native": "^10.1.1",
     "@types/crypto-js": "^4.1.1",
+    "@types/flat": "^5.0.2",
+    "@types/glob": "^7.2.0",
     "@types/i18n-js": "^3.8.2",
-    "@types/jest": "^28.1.4",
+    "@types/jest": "^28.1.6",
     "@types/lodash": "^4.14.182",
+    "@types/papaparse": "^5.3.2",
     "@types/react": "^17.0.39",
-    "@types/react-native": "^0.69.2",
+    "@types/react-native": "^0.69.3",
     "@types/react-test-renderer": "^18.0.0",
-    "@typescript-eslint/eslint-plugin": "^5.30.5",
-    "@typescript-eslint/parser": "^5.30.5",
+    "@types/source-map": "^0.5.7",
+    "@typescript-eslint/eslint-plugin": "^5.30.6",
+    "@typescript-eslint/parser": "^5.30.6",
     "@withshepherd/faker": "^5.5.5",
-    "babel-jest": "^28.1.2",
+    "babel-jest": "^28.1.3",
     "babel-loader": "^8.2.5",
     "babel-plugin-dotenv-import": "^2.2.0",
     "babel-plugin-macros": "^3.0.1",
     "babel-plugin-module-resolver": "^4.1.0",
-    "camelcase": "^6.3.0",
-    "eslint": "^8.19.0",
+    "camelcase": "6.3.0",
+    "eslint": "^8.20.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-prettier": "^8.5.0",
     "eslint-import-resolver-custom-alias": "^1.3.0",
     "eslint-plugin-flowtype": "^8.0.3",
     "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jest": "^26.5.3",
+    "eslint-plugin-jest": "^26.6.0",
     "eslint-plugin-jsx-a11y": "^6.6.0",
     "firebase": "^9.9.0",
     "firebase-admin": "^11.0.0",
     "flat": "^5.0.2",
     "husky": "^8.0.1",
-    "jest": "^28.1.2",
+    "jest": "^28.1.3",
     "jest-fetch-mock": "^3.0.3",
     "jsonschema": "^1.4.1",
     "lint-staged": "^13.0.3",
